backend/requirements.txt
# UDC Polaris Backend - Python Requirements
# Python 3.11+ required

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Multi-Agent Framework
crewai>=0.1.0
langchain>=0.1.0
langchain-anthropic>=0.1.0

# LLM Providers
anthropic>=0.7.0
openai>=1.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Vector Database
chromadb>=0.4.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0

# Advanced Embeddings
sentence-transformers>=2.2.2

# PDF Generation
weasyprint>=60.0
jinja2>=3.1.0

# Web Scraping (for PropertyShop if needed)
beautifulsoup4>=4.12.0
scrapy>=2.11.0
requests>=2.31.0

# API & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Monitoring & Logging
sentry-sdk[fastapi]>=1.38.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0
faker>=20.0.0

# Code Quality
black>=23.11.0
ruff>=0.1.5
mypy>=1.7.0
types-requests>=2.31.0
types-redis>=4.6.0

# Development
ipython>=8.17.0

